arm boards: Set `irqchip.gicv3_pseudo_nmi=1` on the kernel command line
In order to get pseudo-NMI enabled, we not only have to enable the
kernel config `CONFIG_ARM64_PSEUDO_NMI` but we _also_ have to set a
kernel command line.

The reason for need both the config and command line is that, for a
long time, pseudo-NMI was considered unstable and also had big
performance implications. The idea was that a Linux distribution could
enable pseudo-NMI support in the kernel without exposing everyone to
instability and performance issues. Individual users who wanted
pseudo-NMI could then turn it on via the kernel command line and
accept the downsides.

Today, pseudo-NMI still has _some_ performance issues but they are
fairly minor. The small decrease in some benchmarks is justified by
the fact that we'll be able to gather much better traces of locked up
systems.

Today, pseudo-NMI appears to be quite stable, though Mark Rutland at
ARM still expresses some worries [1] [2] we are not aware of any
concrete problems and believe the feature to be stable.

[1] https://lore.kernel.org/lkml/ZFvGqD%2F%2Fpm%2FlZb+p@FVFF77S0Q05N.cambridge.arm.com/
[2] https://lore.kernel.org/lkml/ZNDKVP2m-iiZCz3v@FVFF77S0Q05N.cambridge.arm.com

A few notes:
- Though this looks like it's turning the config on for every ARM
  board, it actually isn't. We've only turned on the kernel config for
  chromeos-5.15+ since we haven't backported all the support to
  previous kernels. All boards are listed here so that they can get
  enabled whenever they uprev.
- For some Mediatek boards, pseudo-NMI won't even be enabled after
  they uprev to a newer kernel despite this kernel command line
  argument and the `CONFIG_ARM64_PSEUDO_NMI` kernel config. This is
  because many older Mediatek boards have a bug in their
  firmware. Those boards are still listed here so that when we uprev
  the firmware of those boards then pseudo-NMI will get enabled. See
  b/281831288
- For oak/elm (mt8173) we don't bother enabling this because those
  boards don't have GICv3 and thus this command line option doesn't
  make sense there.

BUG=b:197061987
TEST=pseudo-NMI works on trogdor and mtk boards w/ new firmware

Cq-Depend: chrome-internal:6542779
Change-Id: I99b47c2b5ea97bd7b0d99857824723f49a0a1b60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4923417
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>