baseboard-asurada: Use TEO cpuidle governor
Match all the other ARM boards and specify this on the kernel command
line. This doesn't matter (yet) for Asurada since it's still on the
chromeos-5.4 and the command line argument isn't needed there, but
this is a good idea for futureproofing. It won't hurt on chromeos-5.4.

BUG=b:241818255
TEST=CQ

Change-Id: I9d65676feada3935acba284b2457989f338f884d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/4937618
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>