strongbad: Set scheduler boost from 60 to 70
This is like CL:4369791 for trogdor, but I managed to not notice that
CL was in overlay-trogdor, not baseboard-trogdor. Sigh. That means we
missed boosting strongbad. Boost it now to match trogdor since there's
no reason for them to be different.

Also note that we were missing `boost-top-app` for strongbad. I don't
_think_ that matters since I believe that was just for ARC++, but
let's at least make it consistent with trogdor.

BUG=b:217825939, b:243336790, b:248028313
TEST=finch

Change-Id: Ifd541833507a4b1d9210b04cafbd8326c1c7e44d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5084479
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>