all x86_64 boards: build the VAAPI decoder for crosvm
This includes (but does not enable) support for crosvm to decode using
VAAPI directly, instead of relying on Chrome to do so. Having support
built-in will enable us to flip it using a command-line argument and
easily test it.

This was enabled for a couple of boards only, but there is no reason to
not make it more widely available.

BUG=b:262824148
TEST=build image

Change-Id: I2d55905e441107a1a09268b626c136ebb7b6254e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5121811
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>