geralt: Disable pseudo NMI in kernel command line
Pseudo-NMI currently doesn't work well with the WIP TEE kernel and/or
firmware and causes kernel hard lockup on boot.
There's also one feedback that Geralt may encounter the hard lockup even
without the TEE kernel. The cause is unknown as of now and MediaTek
doesn't have enough resource to investigate.

Temporarily disable pseudo-NMI to unblock the development, but we need
to work with MTK to re-investigate in the future.

BUG=b:316517822
TEST=geralt can boot with local built kernel that has TEE support

Change-Id: I08d7de217f206eb5c1fc506825b399775d5b7634
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5130471
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>