overlay-brya: use kernel 5.15
While running the public ChromiumOS images in KernelCI we noticed some
tests were failing and traced it back to using an older version of the
kernel than the version used in private ChromeOS builds. This change
sets the kernel version used in the public overlay for brya to the
version currently set in the corresponding private overlay.

TEST=local builds, KernelCI builds; CQ.

Change-Id: I9be25954641606670f6ffac73198151d5393a642
Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5129467
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Terry Cheong <htcheong@chromium.org>