overlay-dedede: use kernel 6.1
While running the public ChromiumOS images in KernelCI we noticed some
tests were failing and traced it back to using an older version of the
kernel than the version used in private ChromeOS builds. This change
sets the kernel version used in the public overlay for dedede to the
version currently set in the corresponding private overlay.

TEST=local builds, KernelCI builds; CQ.

Change-Id: I44ebddfd7ec769b58950e178384c94dabf85e522
Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5129468
Reviewed-by: Mike Frysinger <vapier@chromium.org>