overlay-zork: use kernel 5.15 and enable TPM2
While running the public ChromiumOS images in KernelCI we noticed some
tests were failing and traced it back to using an older version of the
kernel than the version used in private ChromeOS builds. This change
sets the kernel version used in the public overlay for zork to the
version currently set in the corresponding private overlay.

Another class of frequent failures is due to missing TPM flags, so we
also add those in the same way they're defined in the corresponding
private overlay.

BUG=b:298548374, b:280452610
TEST=local builds, KernelCI builds; CQ.

Change-Id: I26f08412cf3fb3c9bc3f189a051aced60fd2bbff
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5129470
Reviewed-by: Raul Rangel <rrangel@chromium.org>