amd64-generic: Update kernel splitconfig name
Sometimes people are confused by the "chromiumos-<arch>" naming. It can
give an impression that it's a common config used by all devices using
<arch>.

That's incorrect. It's "<arch>/common.config" that serves the purpose.
Such confusion tend to happen in cases like when the name is brought up
in a context that doesn't also mention common.config.

To mitigate that, we proposed to update the naming convention to
"chromiumos-<arch>-generic" in all CrOS kernel repos.

Update the kernel splitconfig name in the overlays to match that.

BUG=b:317315065
TEST=PKG=(chromeos-kernel-{4_{14,19},5_{4,10,15},6_{1,6},upstream})
     for pkg in "${PKG[@]}"; do
        emerge-amd64-generic -j $pkg || break;
     done

Cq-Depend: chromium:5175051,chromium:5175071,chromium:5175091
Cq-Depend: chromium:5175111,chromium:5171951,chromium:5171469
Cq-Depend: chromium:5171955
Change-Id: I51ffa81cedb08e6fae4d07a78b71e649fffd690e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5151755
Reviewed-by: Knox Chiou <knoxchiou@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>