chipset-jsl: Make GuC/HuC/DMC firmware builtin
As part of upreving dedede to 6.1, it was noticed that the DMC firmware
fails to load. This is because the .bin is not built-in and the i915
kernel driver no longer waits for the rootfs to be ready before
attempting to load anything.

Make the DMC firmware built-in, and clean up the flags so that only one
flag is needed to include JSL/i915 related firmware.

BUG=b:316622744
TEST=-Build and load kernel on dedede and dedede-kernelnext boards
     -Inspect dmesg logs for successful fw load
     -Verify .bins removed from /lib/firmware
     -`tast run <dut-ip> graphics.IgtKms.kms_atomic`

Cq-Depend: chromium:5153264
Change-Id: Ie007a504f3ccbc1f23725d68af85627c3fe6eab8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5152178
Tested-by: Mark Hasemeyer <markhas@google.com>
Commit-Queue: Mark Hasemeyer <markhas@google.com>
Reviewed-by: Ivan Chen <yulunchen@google.com>
Reviewed-by: Tim Van Patten <timvp@google.com>