chipset-mt8188g: add +crypto to RUSTFLAGS
Since we build C and C++ with `+crypto`, we can do that with Rust, too.
Rust doesn't support `+crypto` (b/316583777), but `+crypto` is just an
alias that expands to these three things anyway.

This also removes +armv8-a from the target features, since passing this
feature leads to a lot of noise when building Rust code; armv8-a is the
default for ARM64 anyway.

BUG=b:318435203
TEST=emerge-geralt memd

Change-Id: Iad7e28b5c5d4fbce3d4e5fda755c7d27db4ee096
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5164311
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>