chipset-qc7180: add +crypto to RUSTFLAGS
Since we build C and C++ with `+crypto`, we can do that with Rust, too.
Rust doesn't support `+crypto` (b/316583777), but `+crypto` is just an
alias that expands to these three things anyway.

This also removes +armv8-a from the target features, since passing this
feature leads to a lot of noise when building Rust code; armv8-a is the
default for ARM64 anyway.

BUG=b:318435203
TEST=emerge-trogdor memd

Change-Id: Ieeb19abbeacb80197bd9e4ed6561bfae6699eda2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5164312
Reviewed-by: Bob Moragues <moragues@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>