chipset-mt8186: add +crypto to RUSTFLAGS
Since we build C and C++ with `+crypto`, we can do that with Rust, too.
Rust doesn't support `+crypto` (b/316583777), but `+crypto` is just an
alias that expands to these three things anyway.

This also removes +armv8-a from the target features, since passing this
feature leads to a lot of noise when building Rust code; armv8-a is the
default for ARM64 anyway.

BUG=b:318435203
TEST=emerge-corsola memd

Change-Id: I7e7f15c65e3d8723656d52ff4c3cd4e8a0ac8d12
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5164313
Reviewed-by: Knox Chiou <knoxchiou@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>