reef: Tolerate missing customization_id in VPD for pyro/sand
While normal devices should always have a customization_id present,
devices in the lab or in the wild could be modified and missing RO_VPD
values.  Tolerate this condition.

Note we are not able to make this change for reef as well, as the
customization_id is necessary to differentiate electro vs. basking.

Note: this is not a complete fix for b/318334523.  While we can make
cros_config tolerate missing VPD values, other OS services are likely
to have issues with an erased VPD.

BUG=b:318334523
TEST=emerge, inspect config json

Change-Id: I70d5d59121066b5527547942ad8ccfed89de523a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5178309
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>