reven: Depend on busybox at buildtime
Busybox is a dependency of the factory and Flexor initramfs
and needs a valid network connection to build (due to downloading
and archive). This fails on builders if invoked as part of build-image.

This patch adds a dependency on busybox for the reven-bsp, circumventing
the problem since busybox will build (and download) as part of a normal
build-packages run. This doesn't fix the underlying issue, but allows
Flexor to be build on the builders for now.

BUG=b:321687359
TEST=Run cros build-packages --board=reven and then
cros build-image --board=reven flexor and observe
busybox not being rebuilt in the build-image step.

Change-Id: I628fa0518adb363f1dce68836b8c16473b18de70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5224562
Reviewed-by: Joseph Sussman <josephsussman@google.com>
Tested-by: Bastian Kersting <bkersting@google.com>
Commit-Queue: Bastian Kersting <bkersting@google.com>