amd64-host: Disable lldb for bazel stage1 builds
I'm trying to prune up the number of packages included with the stage2
SDK, and also flatten the dep graph. `lldb` pulls in a few dependencies
that we can avoid if we disable lldb. In the future we can have an
lldb ebuild that produces the host tool for interactive use. For now,
we don't need lldb to build any packages.

BUG=b:316588646
TEST=BOARD=amd64-generic bazel build --config=hash_tracer @portage//host/sys-devel/llvm

Cq-Depend: chromium:5235403
Change-Id: Idddb4079fe8539e979afc0ffab132ad3e5f36394
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5235544
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>