chipset-stnyridge: clarify kernel version is 5.10
As far as I can tell, the only thing that inherits from stnyridge is
grunt, which was already updated to 5.10, but the chipset default USE
flags still selected 4.14.

Changing the default in chipset-stnyridge means the override in
overlay-grunt can be removed, and it makes it more obvious that devices
with this chipset are no longer using Linux 4.14.

This should have no actual effect on the build.

BUG=b:178032896
TEST=cros build-packages --board=grunt

Change-Id: I7d64a35a8ba0d0b12cccf9f5386f7168c7309392
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5235545
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Matt Davis <mattedavis@google.com>