Add Bluetooth sysprop for boards using QCA BT chip
This adds a sysprop to force Floss to use scan parameters that are known
to be good in QCA chip.

Some models that using RTK controller might also get affected but since
the parameters are used in BlueZ, the performance should be about the
same.

The affected boards include, guybrush, grunt, jacuzzi, kukui, trogdor
and strongbad.

BUG=b:320223219
TEST=emerge-guybrush chromeos-base/chromeos-bsp-guybrush, deploy and
check the file in installed in /etc/bluetooth

Change-Id: I78f45608f30a928113836370ee61915ac1fff028
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5232628
Reviewed-by: Ying Hsu <yinghsu@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: Yun-hao Chung <howardchung@chromium.org>
Reviewed-by: Ashley Shen <shenashley@google.com>
Commit-Queue: Yun-hao Chung <howardchung@chromium.org>