baseboard-trogdor: Use IPA as governor for CPU thermal zones
With recent upstream kernels IPA is profoundly unhappy when it is
used for thermal zones that don't have the two regular trip points
it expects. That makes it unsuitable for being used as default
governor as we did until now. We are switching to stepwise as
default governor, therefore we need to configure the CPU zones to
use IPA, to keep things (mostly) as they were.

BUG=b:325969096
TEST=on lazor:
  grep power_allocator /sys/class/thermal/thermal_zone*/policy | wc -l
    => 10

Cq-Depend: chromium:5315432
Change-Id: I94feef80f6a76c8c911e7f1983c641b20fd85ef7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5315976
Reviewed-by: Bob Moragues <moragues@chromium.org>
Commit-Queue: Bob Moragues <moragues@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>