modemfwd-helpers: Move zork modem firmware into squashfs
Due to the firmware directory structure of some new modems, we cannot
individually compress modem firmware files. This patch moves all
firmware into a compressed squashfs to save substantial space, and
mounts the squashfs in the same location as it was previously,
making this change transparent to firmware update daemons, our tests,
etc.

For zork, this change saves 27.2MiB, a 24% savings.

This change was tested with ModemFWManifestVerification on Zork.

BUG=b:213369060,b:325679411
TEST=Performed above validation

Change-Id: I937ecb955d16e0e7f85884ca6f732d465388e0d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5341479
Tested-by: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>