modemfwd-helpers: Move brya modem firmware into squashfs
Due to the firmware directory structure of some new modems, we cannot
individually compress modem firmware files. This patch moves all
firmware into a compressed squashfs to save substantial space, and
mounts the squashfs in the same location as it was previously,
making this change transparent to firmware update daemons, our tests,
etc.

For brya, this change saves 59.2MiB, a 32% savings.

This change was tested with ModemFWManifestVerification on Brya.

BUG=b:213369060,b:325679411
TEST=Performed above validation

Change-Id: I3d3bf29246ab424f051a35cc1a5466f6a0b5d9cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5341480
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>