modemfwd-helpers: Move rex modem firmware into squashfs
Due to the firmware directory structure of some new modems, we cannot
individually compress modem firmware files. This patch moves all
firmware into a compressed squashfs to save substantial space, and
mounts the squashfs in the same location as it was previously,
making this change transparent to firmware update daemons, our tests,
etc.

For rex, this change saves 65.6MiB, a 38% savings.

This change was tested with ModemFWManifestVerification on Rex.

BUG=b:213369060,b:325679411
TEST=Performed above validation

Change-Id: I0b301621f2358cc45637f0505c42475deaada185
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5341481
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>