geralt: Enable kernel serial console
For some reason OPTEE logging requires kernel serial console to be on,
or system will crash.

To mitigate the impact and allow development on top of latest Geralt
kernel tree, enable the serial console temporarily with loglevel set to
0 i.e. the console will be turned on but being silent.

This should be removed once the issue with OPTEE logging is resolved.

BUG=b:329756745
TEST=CQ

Change-Id: Iffca6793c8425ecf855d80d5bed09aa575d3743a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5371391
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Fei Shao <fshao@chromium.org>