karis: workaround fix for DSP error after resume
As mentioned in the bug we found that DSP error is no longer reproduced
while not setting param blobs to both DTS and EQIIR on DSP, although the
root cause is not clarified yet.

Upon that, the workaround fix is proposed to apply EQIIR on CRAS in
place of DSP. The replacement is functonally equivalent.

BUG=b:328536888
TEST=stress 200 times for audio playback and system suspend/resume

Change-Id: I6e17db31ab60466c150bd487c9047f10802948bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5379870
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>