overlays: add legacy_usb_passthrough USE flag
For permissive USB passthrough (see go/lofn) we will need to default to
legacy behavior for platforms which do not (and likely will not) have
USB annotations. These 3 platforms are also excluded in
https://source.chromium.org/chromiumos/chromiumos/codesearch/+/main:src/platform/tast-tests/src/go.chromium.org/tast-tests/cros/remote/bundles/cros/firmware/usb_annotations.go;l=87;drc=b831391f486f3f8c9a7306487f6ec36862d1be85,
and we therefore feel that since this test is passing for all platforms
this should be a pretty 'authoritative' list. oak does not appear to
have an overlay anymore as it is likely too old.

BUG=b:267951284
TEST=unit test

Change-Id: I356f7453a954b5402bd1be0a5600f05c3a407646
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5426651
Commit-Queue: maciek swiech <drmasquatch@google.com>
Reviewed-by: Shawn Ku <shawnku@chromium.org>
Tested-by: maciek swiech <drmasquatch@google.com>