geralt: sys-kernel: Hotfix for sof suspend issue
This fix is not fully reviewed yet, but if without it the factory can't
continue testing during building.

Add this to work around the issue and unblock factory.

Remove this once it's in 6.1 and Geralt ToT picks that up.

BUG=b:331302167
TEST=emerge-geralt chromeos-kernel-6_1

Change-Id: I83a395b2cdd8b3ee8995c4957db436ab9d4e370c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/5446259
Auto-Submit: Fei Shao <fshao@chromium.org>
Reviewed-by: Knox Chiou <knoxchiou@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Knox Chiou <knoxchiou@chromium.org>