chipset-mt8196: Disable NaCl
We have announced the deprecation of NaCl. It will be disabled since
M132 while remain support until M138.
For the device in development, it should be safe to remove NaCl from
now on.
Disable NaCl on MT8196 so that we can remove the ARMv7
toolchain and enable ARMv9 features.

BUG=b:384407235
TEST=CQ

Change-Id: If0b51ae98b6773e9bb2d75245b795dc67b3d1376
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6096420
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Yu-Che Cheng <giver@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Tested-by: Yu-Che Cheng <giver@chromium.org>