trulo: Pin ISH firmware
Pin the ISH firmware for trulo builds. For each project with an ISH
target, we will query the configs for the ISH version and unpack it.
This all depends on a new host config command 'get-firmware-version'
and a USE flag zephyr_ish_pinned (which will be set by default for
nissa).

BUG=b:389126940, b:380006481
TEST=emerge-nissa chromeos-base/chromeos-bsp-nissa

Cq-Depend: chromium:6293605
Change-Id: I8de1f704adf2e0689cc9f025d74e0772bb3d33e3

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6230066
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Lalith Rajendran <lalithkraj@google.com>