amd64-host: Add riscv32-cros-elf toolchain
In order for the toolchain to be part of the SDK it needs to be in one
of the toolchain.conf files in the overlays.

This toolchain is currently used for building FPMCU firmware. The FPMCU
firmware is independent of the CrOS board, so we put the toolchain in
amd64-host.

BUG=b:355044350
TEST=sudo `which cros_setup_toolchains` -t riscv32-cros-elf --nousepkg

Change-Id: Id699895e855ff845c205659e87c7463ffe3127e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6300631
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>