baseboard-rauru: Bump up size of stateful partition
In preparation for using dev_image.block for developer tools
bump up size of stateful partition on rauru. This is due
to the fact that the filesystem metadata for dev_image.block
increases the utilization.

Note: this only impacts the size of images; on installation,
the partition will expand to occupy the size of disk.

BUG=b:398713190
TEST=CQ

Change-Id: Ib2f6069a3017adfe8c9523ef0b1b3256e86a39ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6302264
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Knox Chiou <knoxchiou@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>