[owners] Remove rekanorman@chromium.org from overlay-nissa/OWNERS
This suggested change is automatically generated based on commit
and review activity to improve the reliability of OWNERS suggestions
and CL review latency.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.
If no CR -1 vote is set after 5 days, the CL will be submitted.
OWNERS expectations for the chromium project are at:
https://chromium.googlesource.com/chromium/src/+/refs/heads/main/docs/code_reviews.md#expectations-of-owners

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/rekanorman@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iae264f085fbb20dcc00df5268329133adad49bd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6410539
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Tested-by: Henry Sun <henrysun@google.com>
Commit-Queue: Henry Sun <henrysun@google.com>
Reviewed-by: Henry Sun <henrysun@google.com>