kukui: set adaptive-charging-hold-delta-percent
Testing on Katsu identified potential limitations
with the EC and oscillations when the hold delta
was set to 0.

Implement a solution similar to that used on the Atlas
project. Specifically, update the
adaptive-charging-hold-delta-percent configuration
value to 3 for all Kukui boards since they use the
same EC.

BUG=b:413506450
TEST=cros_sdk emerge-kukui chromeos-config-bsp chroemos-config
and check pref and check charging behavior

Change-Id: I439284940195a5b03eac9f1bc9d9eede15699685
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6499052
Commit-Queue: Jing Li <jingmuli@google.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Kyle Lin <kylelinck@google.com>
Tested-by: Jing Li <jingmuli@google.com>