zork: Disable AMD TEE module
Disable it using initcall_backlist kernel boot argument.
It can't be disabled in Kconfig since that is shared between
several AMD boards.
The driver landed in tree between 5.4 and 5.15.
During the uprev to the 5.15 it was discovered that with it enabled
some new DRM codecs were detected by avtest_labvel_detect.
This in turns caused tast.video.Capability test to fail, as it
detected more codes than expected.
We don't really want to enable new features during kernel uprev,
especially considering that the hardware itself might not be fast
enough to support it flawlessly.
Because of that this functionality was disabled in the kernel.

BUG=b:247635516
TEST=emerge-zork chromeos-kernel-6_6; update_kernel.sh --board=zork; cat /proc/cmdline

Change-Id: I9e1abd8c453afc1bb041152da3e62abf1d8cc984
Signed-off-by: Helen Koike <koike@igalia.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6494764
Reviewed-by: YH Lin <yueherngl@chromium.org>
Reviewed-by: Matt Davis <mattedavis@google.com>
Tested-by: Matt Davis <mattedavis@google.com>
Commit-Queue: Matt Davis <mattedavis@google.com>