reven: Keep AMT alive on the reven\nuc13
AMT often stops responding to network requests on the nuc13, but we
never see the issue on the nuc11. For b/338368444 we disabled power
saving mode for the nuc11's communication controller using a udev rule.

Let's speculatively apply the same fix for the nuc13's communication
controller, to see if we can get the issue on that model to go away.

If the fix ends up being ineffictive we can revert: the only negative
consequence I can think of is other devices with the same comminication
controller may experience increased battery usage in the case where they
are not a desktop, but they would also likely suffer from the same
issue if we kept power saving mode enabled.

BUG=b:425721648
TEST=`build-packages --board reven && build-image --board reven test`
TEST=`cros flash $DUT xbuddy://local/reven/latest/test`
TEST=SSH to the DUT then verify that the ouput of the following is `on`:
TEST=`cat /sys/bus/pci/devices/0000\:00\:16.0/power/control`

Change-Id: Ie99222f7e2fc2b356d646b899da38e859de098f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6652455
Commit-Queue: Nehemiah Dureus <ndureus@chromium.org>
Auto-Submit: Joseph Sussman <josephsussman@google.com>
Reviewed-by: Nehemiah Dureus <ndureus@chromium.org>
Tested-by: Joseph Sussman <josephsussman@google.com>