nami: set `ecc_size` to 16 bytes for ECC for ramoops
The default value of `ecc_size`, the size in bytes of parity date in ECC
for ramoops, is 4 bytes. This change sets `ecc_size` to 16 bytes for
nami, as nami devices experience more severe corruption in ramoops.

In cases where `chromeos_pstore.ecc_size` already exists in the command
line parameter, it is overwritten instead of appending a new
entry. This change ensures that the parameter is only defined once in
the kernel's command line.

BUG=b:413292019
TEST=build image for nami and check config.txt
Change-Id: Iabf1111ced92755e1d5eff6be842460fd46f6ed7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6688394
Commit-Queue: Naoya Tezuka <naoyatezuka@google.com>
Tested-by: Naoya Tezuka <naoyatezuka@google.com>
Reviewed-by: Ashley Shen <shenashley@google.com>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>