screebo: Disable zero-length scroll suppression property
The default value for the flag that suppresses zero scroll events is
enabled for all boards. This is prohibiting the gesture library from
correctly evaluating when users have kept their fingers completely
stationary and is adding an additional non-zero fling when they release
their fingers.

This change disables the flag for the screebo models by creating a new
conf file specifically for rex.

BUG=b:378586077
TEST=unittests, manual testing, creating new touchpad test to verify
this behaviour as well.

Change-Id: I0ead1f0e0d333aaf574e4f50df196bb1c8ca7adb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/6692511
Commit-Queue: Sairah Amuthan <sairahamuthan@google.com>
Reviewed-by: Samuel Jacob <samjaco@google.com>
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: William Mahon <wmahon@google.com>
Tested-by: Sairah Amuthan <sairahamuthan@google.com>