nissa: kaladin/kelsier: Fix speaker dsp name in UCM config
The speaker audio pipeline was not being created on kaladin and kelsier
boards due to a mismatch in the DSP name. The `DspName` in the UCM
`HiFi.conf` files was specified as "Speaker_eq", but the CRAS DSP
expected "speaker_eq" as defined in `dsp.ini`.

This resulted in the speaker EQ not being applied.

BUG=b:461712898
TEST=cras_test_client --dump_dsp

Change-Id: I81fdbd821a04106ed7b42ce91457adc4008f7e4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/board-overlays/+/7171900
Commit-Queue: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Baili Deng <bailideng@google.com>
Tested-by: Terry Cheong <htcheong@chromium.org>